pyproject.toml
[project]
name = "quas"
version = "0.31.0"
description = "Steganography and cryptanalysis CLI toolkit"
readme = "README.md"
authors = [
    { name = "yuyy", email = "ui.zeyu@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "click>=8.3.1",
    "gmpy2>=2.2.2",
    "gmpy2-stubs>=2.2.1.5",
    "kitcat>=1.2.1",
    "magika>=1.0.2",
    "matplotlib>=3.10.8",
    "numpy>=2.3",
    "pikepdf>=10.3.0",
    "pillow>=12.1.0",
    "pycryptodomex>=3.23.0",
    "pyparsing>=3.3.2",
    "python-magic>=0.4.27",
    "random2>=1.0.2",
    "rich>=14.3.2",
    "scipy>=1.17.0",
    "soundfile>=0.13.1",
    "textual-image>=0.8.5",
    "toolz>=1.1.0",
]

[project.scripts]
quas = "quas:main"

[tool.ruff]
target-version = "py314"

[tool.ruff.lint]
ignore = ["E501"]
select = ["E", "F", "UP", "B", "SIM", "I"]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "nuitka[onefile]>=4.0.1",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
